# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Public Barça News API
 * OpenAPI spec version: 0.3.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type NewsStatus = typeof NewsStatus[keyof typeof NewsStatus];


export const NewsStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export interface News {
  id: number;
  slug: string;
  titleAr: string;
  summaryAr: string;
  bodyAr: string;
  /** @nullable */
  coverImagePath: string | null;
  /** @nullable */
  imageAltAr: string | null;
  categoryId: number;
  categoryNameAr: string;
  /** @nullable */
  authorId: number | null;
  /** @nullable */
  relatedMatchId: number | null;
  status: NewsStatus;
  isFeatured: boolean;
  isBreaking: boolean;
  /** @nullable */
  readingMinutes: number | null;
  tags: string[];
  /** @nullable */
  publishedAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AdminDashboard {
  publishedCount: number;
  draftCount: number;
  breakingCount: number;
  featuredCount: number;
  latestUpdated: News[];
}

export interface AdminNewsCreateRequest {
  /** @minLength 1 */
  titleAr: string;
  /** @minLength 1 */
  summaryAr: string;
  /** @minLength 1 */
  bodyAr: string;
  categoryId: number;
  /** @nullable */
  coverImagePath: string | null;
  /** @nullable */
  imageAltAr: string | null;
  tags: string[];
  isFeatured: boolean;
  isBreaking: boolean;
  /** @nullable */
  readingMinutes: number | null;
}

export type AdminNewsUpdateRequest = AdminNewsCreateRequest;

export interface AdminCategoryCreateRequest {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  nameAr: string;
  /** @nullable */
  nameEn: string | null;
  sortOrder: number;
  isActive: boolean;
}

export type AdminCategoryUpdateRequest = AdminCategoryCreateRequest;

export type AdminTransferRequestStatus = typeof AdminTransferRequestStatus[keyof typeof AdminTransferRequestStatus];


export const AdminTransferRequestStatus = {
  completed: 'completed',
  negotiation: 'negotiation',
  rumor: 'rumor',
} as const;

export interface AdminTransferRequest {
  /** @nullable */
  playerId: number | null;
  /** @minLength 1 */
  playerNameAr: string;
  /** @nullable */
  position: string | null;
  /** @nullable */
  fromClub: string | null;
  /** @minLength 1 */
  toClub: string;
  /** @nullable */
  feeAmount: number | null;
  /** @nullable */
  feeCurrency: string | null;
  /** @nullable */
  feeLabelAr: string | null;
  status: AdminTransferRequestStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  /** @nullable */
  notesAr: string | null;
  /** @nullable */
  announcedAt: string | null;
}

export type AdminMatchRequestStatus = typeof AdminMatchRequestStatus[keyof typeof AdminMatchRequestStatus];


export const AdminMatchRequestStatus = {
  scheduled: 'scheduled',
  live: 'live',
  finished: 'finished',
  postponed: 'postponed',
  cancelled: 'cancelled',
} as const;

export interface AdminMatchRequest {
  /** @minLength 1 */
  competition: string;
  /** @minLength 1 */
  homeTeam: string;
  /** @minLength 1 */
  awayTeam: string;
  /** @nullable */
  homeShort: string | null;
  /** @nullable */
  awayShort: string | null;
  kickoffAt: string;
  /** @nullable */
  venue: string | null;
  status: AdminMatchRequestStatus;
  /** @nullable */
  homeScore: number | null;
  /** @nullable */
  awayScore: number | null;
}

export interface Category {
  id: number;
  slug: string;
  nameAr: string;
  /** @nullable */
  nameEn: string | null;
  sortOrder: number;
  isActive: boolean;
}

export interface NewsListResponse {
  items: News[];
  page: number;
  pageSize: number;
  total: number;
}

export interface Player {
  id: number;
  slug: string;
  nameAr: string;
  /** @nullable */
  nameLatin: string | null;
  position: string;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  noteAr: string | null;
  isActive: boolean;
}

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  scheduled: 'scheduled',
  live: 'live',
  finished: 'finished',
  postponed: 'postponed',
  cancelled: 'cancelled',
} as const;

export interface Match {
  id: number;
  competition: string;
  homeTeam: string;
  awayTeam: string;
  /** @nullable */
  homeShort: string | null;
  /** @nullable */
  awayShort: string | null;
  kickoffAt: string;
  /** @nullable */
  venue: string | null;
  status: MatchStatus;
  /** @nullable */
  homeScore: number | null;
  /** @nullable */
  awayScore: number | null;
}

export type TransferStatus = typeof TransferStatus[keyof typeof TransferStatus];


export const TransferStatus = {
  completed: 'completed',
  negotiation: 'negotiation',
  rumor: 'rumor',
} as const;

export interface Transfer {
  id: number;
  /** @nullable */
  playerId: number | null;
  playerNameAr: string;
  /** @nullable */
  position: string | null;
  /** @nullable */
  fromClub: string | null;
  toClub: string;
  /** @nullable */
  feeAmount: number | null;
  /** @nullable */
  feeCurrency: string | null;
  /** @nullable */
  feeLabelAr: string | null;
  status: TransferStatus;
  confidence: number;
  /** @nullable */
  notesAr: string | null;
  /** @nullable */
  announcedAt: string | null;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Temporary development admin guard rejected the request
 */
export type ForbiddenResponse = ErrorResponse;

export type ListNewsParams = {
search?: string;
category?: string;
featured?: boolean;
breaking?: boolean;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 50
 */
pageSize?: number;
};

export type ListPlayersParams = {
search?: string;
/**
 * @minimum 1
 * @maximum 50
 */
pageSize?: number;
};

export type ListMatchesParams = {
scope?: ListMatchesScope;
};

export type ListMatchesScope = typeof ListMatchesScope[keyof typeof ListMatchesScope];


export const ListMatchesScope = {
  upcoming: 'upcoming',
  previous: 'previous',
} as const;

export type ListTransfersParams = {
status?: ListTransfersStatus;
};

export type ListTransfersStatus = typeof ListTransfersStatus[keyof typeof ListTransfersStatus];


export const ListTransfersStatus = {
  completed: 'completed',
  negotiation: 'negotiation',
  rumor: 'rumor',
} as const;

export type ListAdminNewsParams = {
search?: string;
category?: string;
status?: ListAdminNewsStatus;
featured?: boolean;
breaking?: boolean;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 50
 */
pageSize?: number;
};

export type ListAdminNewsStatus = typeof ListAdminNewsStatus[keyof typeof ListAdminNewsStatus];


export const ListAdminNewsStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;