# lib/api-zod/src/generated/types/news.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Public Barça News API
 * OpenAPI spec version: 0.3.0
 */
import type { NewsStatus } from './newsStatus';

export interface News {
  id: number;
  slug: string;
  titleAr: string;
  summaryAr: string;
  bodyAr: string;
  /** @nullable */
  coverImagePath: string | null;
  /** @nullable */
  imageAltAr: string | null;
  categoryId: number;
  categoryNameAr: string;
  /** @nullable */
  authorId: number | null;
  /** @nullable */
  relatedMatchId: number | null;
  status: NewsStatus;
  isFeatured: boolean;
  isBreaking: boolean;
  /** @nullable */
  readingMinutes: number | null;
  tags: string[];
  /** @nullable */
  publishedAt: Date | null;
  createdAt: Date;
  updatedAt: Date;
}