# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Public Barça News API
 * OpenAPI spec version: 0.3.0
 */

export interface Player {
  id: number;
  slug: string;
  nameAr: string;
  /** @nullable */
  nameLatin: string | null;
  position: string;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  noteAr: string | null;
  isActive: boolean;
}