# lib/api-zod/src/generated/types/transfer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Public Barça News API
 * OpenAPI spec version: 0.3.0
 */
import type { TransferStatus } from './transferStatus';

export interface Transfer {
  id: number;
  /** @nullable */
  playerId: number | null;
  playerNameAr: string;
  /** @nullable */
  position: string | null;
  /** @nullable */
  fromClub: string | null;
  toClub: string;
  /** @nullable */
  feeAmount: number | null;
  /** @nullable */
  feeCurrency: string | null;
  /** @nullable */
  feeLabelAr: string | null;
  status: TransferStatus;
  confidence: number;
  /** @nullable */
  notesAr: string | null;
  /** @nullable */
  announcedAt: Date | null;
}